+++ SigebiWeb/src/main/java/com/sigebi/service/impl/PersonasServiceImpl.java
@@ -2,13 +2,17 @@ package com.sigebi.service.impl;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 import javax.persistence.criteria.Predicate;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
@@ -28,6 +32,8 @@ import com.sigebi.service.DepartamentosService;
 import com.sigebi.service.DependenciasService;
 import com.sigebi.service.EstamentosService;
 import com.sigebi.service.PersonasService;
+import com.sigebi.service.UtilesService;
+import com.sigebi.util.exceptions.SigebiException;
 
 
 @Service
@@ -43,6 +49,8 @@ public class PersonasServiceImpl implements PersonasService{
 	private DependenciasService dependenciasService;
 	@Autowired
 	private EstamentosService estamentosService;
+	@Autowired
+	private UtilesService utiles;
 	
 	public PersonasServiceImpl(IPersonasDao personasDao) {
         this.personasDao = personasDao;
@@ -50,44 +58,120 @@ public class PersonasServiceImpl implements PersonasService{
 	
 	@Override
 	@Transactional(readOnly = true)
-	public List<Personas> findAll() {
-		return (List<Personas>) personasDao.findAll();
+	public List<Personas> listar() throws SigebiException {
+		List<Personas> personasList = personasDao.findAll();
+		
+		if( personasList.isEmpty()) {
+			throw new SigebiException.DataNotFound("No se encontraron datos");
+		}
+		
+		return personasList;
 	}
 
 	@Override
 	@Transactional(readOnly = true)
-	public Personas findById(int id) {
-		return personasDao.findById(id).orElse(null);
+	public Personas obtener(int id) throws SigebiException {
+		
+		Personas persona = personasDao.findById(id).orElse(null);;
+		
+		if ( persona != null ) {
+			if( persona.getDepartamentos() == null ) {
+				persona.setDepartamentos(new Departamentos());
+			}
+			if( persona.getDependencias() == null ) {
+				persona.setDependencias(new Dependencias());
+			}
+			if( persona.getCarreras() == null ) {
+				persona.setCarreras(new Carreras());
+			}
+			if( persona.getEstamentos() == null ) {
+				persona.setEstamentos(new Estamentos());
+			}
+		}
+		
+		if( persona == null ) {						
+			String mensaje = "La persona con ID: ".concat(String.valueOf(id).concat(" no existe en la base de datos!"));
+			throw new SigebiException.DataNotFound(mensaje);
+		}
+		return persona;
 	}
 
 	@Override
 	@Transactional
-	public Personas save(Personas persona) throws Exception {
-		if( persona.getCarreras() != null) {
+	public Personas guardar(Personas persona) throws SigebiException {
+		
+		if( persona.getCarreras() != null && persona.getCarreras().getCarreraId() != null) {
 			Carreras carrera = carrerasService.findById(persona.getCarreras().getCarreraId());
 			if(carrera == null) {
-				throw new Exception("No se encontro carrera con id: " + persona.getCarreras().getCarreraId());
+				throw new SigebiException.DataNotFound("No se encontro carrera con id: " + persona.getCarreras().getCarreraId());
 			}
 			persona.setCarreras(carrera);
 		}
-		if( persona.getDepartamentos() != null) {
+		if( persona.getDepartamentos() != null && persona.getDepartamentos().getDepartamentoId() != null) {
 			Departamentos departamento = departamentosService.findById(persona.getDepartamentos().getDepartamentoId());
 			if(departamento == null) {
-				throw new Exception("No se encontro departamento con id: " + persona.getDepartamentos().getDepartamentoId());
+				throw new SigebiException.DataNotFound("No se encontro departamento con id: " + persona.getDepartamentos().getDepartamentoId());
 			}
 			persona.setDepartamentos(departamento);
 		}
-		if( persona.getDependencias() != null) {
+		if( persona.getDependencias() != null && persona.getDependencias().getDependenciaId() != null) {
 			Dependencias dependencia = dependenciasService.findById(persona.getDependencias().getDependenciaId());
 			if(dependencia == null) {
-				throw new Exception("No se encontro dependencia con id: " + persona.getDependencias().getDependenciaId());
+				throw new SigebiException.DataNotFound("No se encontro dependencia con id: " + persona.getDependencias().getDependenciaId());
 			}
 			persona.setDependencias(dependencia);
 		}
-		if( persona.getEstamentos() != null) {
+		if( persona.getEstamentos() != null && persona.getEstamentos().getEstamentoId() != null) {
 			Estamentos estamento = estamentosService.findById(persona.getEstamentos().getEstamentoId());
 			if(estamento == null) {
-				throw new Exception("No se encontro estamento con id: " + persona.getEstamentos().getEstamentoId());
+				throw new SigebiException.DataNotFound("No se encontro estamento con id: " + persona.getEstamentos().getEstamentoId());
+			}
+			persona.setEstamentos(estamento);
+		}
+		return personasDao.save(persona);
+	}
+	
+	@Override
+	@Transactional
+	public Personas actualizar(Personas persona) throws SigebiException {
+		
+		if ( persona.getPersonaId() == null ) {
+			throw new SigebiException.DataNotFound("Error: persona id es requerido ");
+		}
+		
+		Personas personaActual = obtener(persona.getPersonaId());
+		
+		if ( personaActual == null ) {
+			String mensaje = "No se pudo editar, la persona ID: "
+					.concat(String.valueOf(persona.getPersonaId()).concat(" no existe en la base de datos!"));
+			throw new SigebiException.DataNotFound(mensaje);
+		}
+		
+		if( persona.getCarreras() != null && persona.getCarreras().getCarreraId() != null) {
+			Carreras carrera = carrerasService.findById(persona.getCarreras().getCarreraId());
+			if(carrera == null) {
+				throw new SigebiException.DataNotFound("No se encontro carrera con id: " + persona.getCarreras().getCarreraId());
+			}
+			persona.setCarreras(carrera);
+		}
+		if( persona.getDepartamentos() != null && persona.getDepartamentos().getDepartamentoId() != null) {
+			Departamentos departamento = departamentosService.findById(persona.getDepartamentos().getDepartamentoId());
+			if(departamento == null) {
+				throw new SigebiException.DataNotFound("No se encontro departamento con id: " + persona.getDepartamentos().getDepartamentoId());
+			}
+			persona.setDepartamentos(departamento);
+		}
+		if( persona.getDependencias() != null && persona.getDependencias().getDependenciaId() != null) {
+			Dependencias dependencia = dependenciasService.findById(persona.getDependencias().getDependenciaId());
+			if(dependencia == null) {
+				throw new SigebiException.DataNotFound("No se encontro dependencia con id: " + persona.getDependencias().getDependenciaId());
+			}
+			persona.setDependencias(dependencia);
+		}
+		if( persona.getEstamentos() != null && persona.getEstamentos().getEstamentoId() != null) {
+			Estamentos estamento = estamentosService.findById(persona.getEstamentos().getEstamentoId());
+			if(estamento == null) {
+				throw new SigebiException.DataNotFound("No se encontro estamento con id: " + persona.getEstamentos().getEstamentoId());
 			}
 			persona.setEstamentos(estamento);
 		}
@@ -96,13 +180,25 @@ public class PersonasServiceImpl implements PersonasService{
 
 	@Override
 	@Transactional
-	public void delete(int id) {
+	public void eliminar(int id) throws SigebiException {
+		
+		if ( utiles.isNullOrBlank(String.valueOf(id)) ) {
+			throw new SigebiException.DataNotFound("Error: persona id es requerido ");
+		}
+		
+		Personas personaActual = obtener(id);
+		
+		if ( personaActual == null ) {
+			String mensaje = "La persona ID: "
+					.concat(String.valueOf(personaActual.getPersonaId()).concat(" no existe en la base de datos!"));
+			throw new SigebiException.DataNotFound(mensaje);
+		}
 		personasDao.deleteById(id);
 	}
 	
 	@Override
-	@Transactional
-	public List<Personas> buscar(Date fromDate, Date toDate, Personas persona, Pageable pageable) {
+	@Transactional(readOnly = true)
+	public List<Personas> buscar(Date fromDate, Date toDate, Personas persona, Pageable pageable) throws DataAccessException {
 		
         List<Personas> personasList = personasDao.findAll((Specification<Personas>) (root, cq, cb) -> {
             Predicate p = cb.conjunction();
@@ -116,10 +212,10 @@ public class PersonasServiceImpl implements PersonasService{
                 p = cb.and(p, cb.equal(root.get("cedula"), persona.getCedula()));
             }
             if (!StringUtils.isEmpty(persona.getNombres())) {
-                p = cb.and(p, cb.like(root.get("nombres"), "%" + persona.getNombres() + "%"));
+                p = cb.and(p, cb.like(cb.lower(root.get("nombres")), "%" + persona.getNombres().toLowerCase() + "%"));
             }
             if (!StringUtils.isEmpty(persona.getApellidos())) {
-                p = cb.and(p, cb.like(root.get("apellidos"), "%" + persona.getApellidos() + "%"));
+                p = cb.and(p, cb.like(cb.lower(root.get("apellidos")), "%" + persona.getApellidos().toLowerCase() + "%"));
             }
             cq.orderBy(cb.desc(root.get("personaId")));
             return p;
